fix: update quantity scheme for inventories and transactions table

+++ database/migrations/2024_04_30_020159_create_inventories_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description')->nullable();
             $table->float('price', 10, 2);
-            $table->integer('quantity');
+            $table->tinyInteger('quantity');
             $table->foreignId('created_by')->constrained('users');
             $table->timestamps();
             $table->softDeletes();
